[#2702] Allow inclusion of defaults in command response messages
The CommandResponseSender interface's sendResponse method has been
extended to require tenant configuration and a registration assertion to
be provided in order to better resemble the TelemetrySender and
EventSender interfaces.

The sender implementations have not (yet) been extended to include the
default properties in the downstream command response in order to keep
the scope of this commit smaller.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>