[#2984] Clean Caches on Device Registry Notifications
The clients for the device registry APIs may register receivers
for notifications of the device registry that are used to remove
objects from the response cache if the corresponding data has
changed.
Protocol adapters check for the existence of a tenant before invoking
the credentials or the registration clients. For performance reasons
the code skips iterating over the registration and credentials caches
and instead just lets the entries expire after their timeout when a
tenant is changed.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>