Use remote container build for creating native image
Running the native-image tool in a local container fails on a rootless
Docker installation because the process is run with a different UID:GID
than the current user's and thus fails to write the created native
executable to the local target folder.

Using a remote container build circumvents this problem by means of
copying the source files and the native executable hence and forth
between the local file system and the container.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>