[#2981] Extend Mongo DB registry to support tenant alias
The Device Registry Management API has been extended to support an
"alias" property in the "Tenant" object. The alias is an optional
alternative identifier that the tenant may be looked up by.

The Mongo DB based Tenant service implementation has been extended to
include the alias when looking up a tenant by ID.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>