Add workflow for testing of native images
Added a GitHub Actions workflow that builds native images for Hono's
Quarkus based components and runs integration tests with them.

The workflow is supposed to be run as a scheduled workflow because it
takes quite long to build the native images on GitHub Actions
infrastructure.

For the time being, the workflow needs to be triggered manually. This is
to gain some experience with it before we schedule it regularly/nightly.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>