Let hono-core module register its classes for reflection
The core module cannot (yet) use the RegisterForReflection annotation
for registering its classes explicitly because the quarkus-core artifact
that the annotation is part of requires Java 11 and hono-core is (still)
at Java 8 level.

However, it still makes sense to put registering the classes for
reflection into the hands of the module that the classes belong to.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>