[#3107] Fix support for classpath extensions in Quarkus images
The configuration of the run-java.sh command used for starting the
Quarkus based components has been changed to no longer invoke the
quarkus-run.jar but to set a main class to execute. This allows for the
additional classpath (including the jars in the extensions folder) to be
considered by the application.

Fixes #3107

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>