[#3083] Only allow mapped target device id in MappedMessage.
Not allowing the whole mapped target address to be set, thereby
making it clear that the tenant identifier can't be overridden.
Also copying of payload data for the Message Mapping is removed
- this shouldn't be necessary anymore (issues were probably due
to the message payload Buffer's ByteBuf getting duplicated to
create a new MqttPublishMessage).
Handling of a mapped payload for a command response message
has been fixed.

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>