[#3163] Do not let AMQP device client send arbitrary properties
Hono's User Guide for the AMQP 1.0 protocol adapter specifies the format
of the messages that a device sends/receives to/from the adapter. The
user guide neither mentions the possibility to receive arbitrary
properties in command messages nor the possibility to include such
properties in the telemetry messages being uploaded by a device.

The amqp-device client module's interfaces do specify methods that
accept arbitrary properties, though. It is unclear what will happen with
the properties in the AMQP adapter so the properties argument has been
removed from the interface methods.

Fixes #3163

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>