Allow sending non-empty commands without content type
The C&C spec does not require a content type to be set when sending
commands. It also allows for commands that contain a payload but do not
specify a content type.

The CommandSender implementations have been adapted accordingly. Also,
the parameter order of the send methods have been changed so that
mandatory parameters come first.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>