Use dedicated module to manage image build
The library modules used by the service modules should not contain the
profiles used for building the container images in order to prevent
unnecessary build steps.

The dedicated hono-services-quarkus module is being used as the parent
of the service modules which allows to selectively do the Quarkus
augmentation and building of the container images.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>