Remove wording regarding custom properties in telemetry messages
The Telemetry API might have been read to expect that custom properties
that an AMQP based device includes in a message that it uploads to the
AMQP adapter are being forwarded in the downstream Telemetry messages.

The wording has been adapted to clarify that this only applies to
properties added by protocol adapters. Thus it depends on the protocol
adapter implementation if and how properties provided by devices are
being included.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>