Allow specifying a heap memory share reserved for GC
When calculating the maximum number of connections supported by a
protocol adapter, the "head room" on the heap that is required for the
garbage collector should be considered as well.

The MemoryBasedConnectionLimitStrategy has been adapted accordingly. The
protocol adapters now support a configuration property for explicitly
setting the percentage of the heap memory not to be used for the live
data-set. The default value is 25%.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>