[#3287] Fix serialization of CommandEndpoint class in native image
The native executable based images of protocol adapters failed to
deserialize a registration assertion if it included command endpoint
information that had been configured for the device.

The CommandEndpoint class has been amended with the
@RegisterForReflection annotation and the registry ITs have been
extended to also verify the transfer of command endpoint info in the
registration assertion.

Fixes #3287

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>