[#3302] Add resource for downloading JWK set
The Authentication Server has been extended with an HTTP resource for
retrieving the set of keys that can be used to validate the signatures
on JWS containing a JWT created by the Auth server.

The Command Router and the Device Registry have been changed to download
the keys from the Auth server during startup and refresh the set
periodically.

Fixes #3302

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>