[#3325] Let binding of Kafka metrics be done via MicrometerRecorder.
Combine Kafka metrics in a MeterBinder bean so that binding of the
metrics to the MeterRegistry is done implicitly via the Quarkus
MicrometerRecorder. This ensures that the binding is done *after*
the MeterFilter beans get added to the registry. This in turn
ensures that the KafkaMetrics are aware of the commonTags
MeterFilters when being bound to the registry.

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>