[#3302] Support setting of alg on validating keys
RFC 8725 (https://datatracker.ietf.org/doc/html/rfc8725) recommends that
applications should make sure that a given key is used with exactly one
signature algorithm only. The validator has been extended with a
configuration property that enables a check which compares the signature
algorithm used in a token with the algorithm that has been specified in
the dynamically retrieved JWK set.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>