[#3302] Always include alg in JWKs
Vert.x' JWK class which is used in the Auth server ITs expects a key's
intended signature algorithm to be included in the JSON that it is
created from. However, if no alg property is included in the JSON, then
it falls back to RS/ES/HS256. This will not work with a JWS that contain
e.g. an ES384 based signature.

The JjwtBasedAuthTokenFactory has therefore been extended to always
include a key's intended signature algorithm in its JWK representation.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>