[#3360] Create native executable command line client
The build jobs have been changed to consistently use the
eclipse/hono-builder container image for building Hono. This image
contains specific versions of Maven, the GraalVM SDK with the
native-image compiler plugin and the UPX tool for compressing binary
artifacts.

The release build job has been extended with an additional invocation
of mvn clean install used to build the native executable of the CLI.

The native executable needs to be built in a separate run because in the
first build the JAR file artifacts are being signed by the Eclipse JAR
signer service which results in signature errors during the native
executable build.

Fixes #3360

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>