[#3390] Allow CLI to send empty messages with no content type
The command line client has been changed to allow sending of messages to
the AMQP adapter that contain no body and no content-type.
It is at the discretion of the AMQP adapter to handle such messages in
an appropriate way. The CLI's error handling has been improved so that
it now always prints the error message received from the AMQP adapter.

Fixes #3390

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>