[#3441] Store DeviceUser in request context objects
The request context objects used by the protocol adapters have been
changed to keep a DeviceUser instance as the authenticated device
instead of a plain Device. This better aligns with the type used in the
CredentialsApiAuthProvider's public methods and also allows for adapters
to use the vert.x User interfaces's methods, e..g for checking if the
user's authentication token has expired.