[#3441] Read tenant from tid claim
As a follow-up on
https://github.com/eclipse-hono/hono/pull/3451#discussion_r1064156006
the JwtAuthHandler has been changed to read the tenant Id from a JWT's
"tid" claim instead of the "iss" claim. This allows for the "iss" claim
to be used for determining where the public key to use for validating
the JWS signature should be looked up from, e.g. the tenant
configuration or device credentials.

Signed-off-by: Kai Hudalla <sophokles.kh@gmail.com>