[#3558] Make HonoKafkaConsumer compatible with Kafka 3.5
The HonoKafkaConsumer has been adapted to explicitly enforce a
rebalance in order to get assigned partitions for auto-created
topics if the metadata.max.age.ms configuration property is not
set or is set to a value greater than 500.

Kafka 3.5 seems to have changed its rebalancing behavior with
regard to auto-created topics. Without this modification,
the assignment of new partitions to the consumer will be deferred
to after the metadata has become stale, which happens after 5 minutes
if not explicitly overridden using the metadata.max.age.ms
configuration property.