[#3569] Use TLS endpoint when connecting to Sandbox
The command line client was still trying to connect to the insecure
ports of the Sandbox.

This has been changed so that the client now uses the TLS endpoints and
requires the user to specify a trust store for validating the server
certificate.

The Getting Started guide has been adapted accordingly.
Also updated the Python Getting Started example script to use Kafka
instead of AMQP 1.0 based messaging infrastructure.