Fixing inconsistent usage of hono.auth
hono.auth had two functions, one where it contained a string to
a) it was used to control the bean behavior in DelegatingAuthenticationServiceProducer
the second function was being
b) the config if that bean

This led to inconsistency during the build process where the string hono.auth. e.g. hono.auth=delegating crashed the application during execution because the app excepts hono.auth to contain an object

Thus, hono.auth controlling DelegatingAuthenticationServiceProducer was renamed to hono.authType